pyproject.toml
[tool.poetry]
name = "lista_de_listas_cli"
version = "0.1.0"
description = ""
authors = ["Gustavo Sousa <ola@gustavosousa.me>"]

[tool.poetry.dependencies]
python = "^3.10"
python-decouple = "^3.6"
SQLAlchemy = "^1.4.32"
click = "^8.0.4"
SQLAlchemy-Utils = "^0.38.2"
pydantic = "^1.9.0"
inquirerpy = "^0.3.3"

[tool.poetry.dev-dependencies]
pytest = "^7.0.1"
blue = "^0.8.0"
isort = "^5.10.1"
prospector = "^1.7.7"
pytest-cov = "^3.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
python_files = ['tests.py', 'test_*.py', '*_tests.py']
python_classes = ["Test", "Acceptance"]
python_functions = ["test"]
testpaths = ["tests"]

[tool.coverage.report]
show_missing = true